Move sfx module into atoms/sfx with dedicated tests directory
Aligns the new SFX utility with the atoms/<name>/tests/<name>.test.js
layout instead of the flat src/services location.
